+++ src/gtp.h
@@ -31,9 +31,9 @@ public:
     MethodMap m_commandMethods;
     void registerMethod(const std::string& cmd_name, MethodMap::mapped_type gcm);
 
-    typedef std::map<std::string, int*> IntParamMap;
+    typedef std::map<std::string, uint*> IntParamMap;
     IntParamMap m_intParams;
-    void registerIntParam(int* v, const std::string& label);
+    void registerIntParam(uint* v, const std::string& label);
 
     typedef std::map<std::string, double*> DoubleParamMap;
     DoubleParamMap m_doubleParams;
@@ -65,16 +65,16 @@ public:
     void clear_interrupt();
 
 private:
-    double m_komi;
+    float m_komi;
     Board m_board;
     volatile bool _needs_interrupt;
 
-    int m_monte_1ply_playouts_per_move;
-    int uct_kPlayouts; //11
-    int uct_kExpandThreshold; //5
-    int uct_kStepSize; //100
+    uint m_monte_1ply_playouts_per_move;
+    uint uct_kPlayouts; //11
+    uint uct_kExpandThreshold; //5
+    uint uct_kStepSize; //100
     double uct_kUctK; //1.0
-    int max_think_millis;
+    uint max_think_millis;
 
     double getMoveValue(BoardState color, Point p);
